refactor(plugin): extract limited/{command,surface_store,pixel_buffer}.rs (#1092)
Final batch of LimitedAccess section extractions; gpu_context/mod.rs
no longer carries any LimitedAccess callback body beyond the
GpuContext-handle clone/drop pair (the vtable-handle ops that retain
the parent file's "GpuContextLimitedAccess vtable" banner).

* limited/command.rs (~525 LoC) — RhiCommandQueue Arc lifecycle +
  create_command_buffer from a queue handle, CommandBuffer drop +
  consume-semantics commits (v7), GpuContextLimitedAccess command-
  queue accessor + create_command_buffer shortcut +
  copy_pixel_buffer_to_texture + blit_copy + the macOS
  blit_copy_iosurface slot.

* limited/surface_store.rs (~120 LoC) — surface_store accessor and
  check_out_surface FD-resolving wrappers.

* limited/pixel_buffer.rs (~335 LoC) — PixelBuffer Arc lifecycle
  (clone, drop, strong_count, plane_base_address, plane_size) plus
  acquire_pixel_buffer / get_pixel_buffer /
  resolve_pixel_buffer_by_surface_id method dispatch. Combines the
  two non-adjacent PB sections from the original file under one
  resource-scope module.

Parent gpu_context/mod.rs's now-unused slice_from_raw +
write_id_bytes + handle_as_gpu_context imports dropped at the same
time (FullAccess code on the way down keeps write_err +
pixel_format_from_raw).

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
