+++ libs/streamlib-engine/src/core/plugin/host_services/gpu_context/limited/mod.rs
@@ -8,7 +8,10 @@
 //! static wires the function pointers up by name.
 
 mod buffer;
+mod command;
 mod escalate;
+mod pixel_buffer;
+mod surface_store;
 mod texture;
 mod texture_registration;
 mod video_source_timeline;
@@ -21,8 +24,25 @@ pub(in crate::core::plugin::host_services) use buffer::{
     host_gpu_lim_drop_index_buffer, host_gpu_lim_drop_storage_buffer,
     host_gpu_lim_drop_uniform_buffer, host_gpu_lim_drop_vertex_buffer,
 };
+pub(in crate::core::plugin::host_services) use command::{
+    host_gpu_lim_blit_copy, host_gpu_lim_blit_copy_iosurface, host_gpu_lim_clone_rhi_command_queue,
+    host_gpu_lim_command_queue, host_gpu_lim_commit_and_wait_command_buffer,
+    host_gpu_lim_commit_command_buffer, host_gpu_lim_copy_pixel_buffer_to_texture,
+    host_gpu_lim_copy_texture_command_buffer, host_gpu_lim_create_command_buffer,
+    host_gpu_lim_create_command_buffer_from_queue, host_gpu_lim_drop_command_buffer,
+    host_gpu_lim_drop_rhi_command_queue,
+};
 pub(in crate::core::plugin::host_services) use escalate::{
     host_gpu_lim_escalate_begin, host_gpu_lim_escalate_end,
+};
+pub(in crate::core::plugin::host_services) use pixel_buffer::{
+    host_gpu_lim_acquire_pixel_buffer, host_gpu_lim_clone_pixel_buffer,
+    host_gpu_lim_drop_pixel_buffer, host_gpu_lim_get_pixel_buffer,
+    host_gpu_lim_plane_base_address_pixel_buffer, host_gpu_lim_plane_size_pixel_buffer,
+    host_gpu_lim_resolve_pixel_buffer_by_surface_id, host_gpu_lim_strong_count_pixel_buffer,
+};
+pub(in crate::core::plugin::host_services) use surface_store::{
+    host_gpu_lim_check_out_surface, host_gpu_lim_surface_store,
 };
 pub(in crate::core::plugin::host_services) use texture::{
     host_gpu_lim_acquire_texture, host_gpu_lim_clone_texture, host_gpu_lim_drop_pooled_texture_handle,
