+++ libs/streamlib-engine/src/core/plugin/host_services/gpu_context/mod.rs
@@ -23,32 +23,43 @@ use streamlib_plugin_abi::{
 
 use super::host_callbacks;
 use super::run_host_extern_c;
-use super::shared::wire::{slice_from_raw, write_err, write_id_bytes};
+use super::shared::wire::write_err;
 
 mod limited;
 mod scope_token;
 mod shared;
 
 use limited::{
-    host_gpu_lim_acquire_index_buffer, host_gpu_lim_acquire_storage_buffer,
-    host_gpu_lim_acquire_texture, host_gpu_lim_acquire_uniform_buffer,
-    host_gpu_lim_acquire_vertex_buffer, host_gpu_lim_clear_video_source_timeline_semaphore,
-    host_gpu_lim_clone_index_buffer, host_gpu_lim_clone_storage_buffer, host_gpu_lim_clone_texture,
+    host_gpu_lim_acquire_index_buffer, host_gpu_lim_acquire_pixel_buffer,
+    host_gpu_lim_acquire_storage_buffer, host_gpu_lim_acquire_texture,
+    host_gpu_lim_acquire_uniform_buffer, host_gpu_lim_acquire_vertex_buffer,
+    host_gpu_lim_blit_copy, host_gpu_lim_blit_copy_iosurface, host_gpu_lim_check_out_surface,
+    host_gpu_lim_clear_video_source_timeline_semaphore, host_gpu_lim_clone_index_buffer,
+    host_gpu_lim_clone_pixel_buffer, host_gpu_lim_clone_rhi_command_queue,
+    host_gpu_lim_clone_storage_buffer, host_gpu_lim_clone_texture,
     host_gpu_lim_clone_texture_registration, host_gpu_lim_clone_uniform_buffer,
-    host_gpu_lim_clone_vertex_buffer, host_gpu_lim_drop_index_buffer,
-    host_gpu_lim_drop_pooled_texture_handle, host_gpu_lim_drop_storage_buffer,
-    host_gpu_lim_drop_texture, host_gpu_lim_drop_texture_registration,
-    host_gpu_lim_drop_uniform_buffer, host_gpu_lim_drop_vertex_buffer, host_gpu_lim_escalate_begin,
-    host_gpu_lim_escalate_end, host_gpu_lim_host_video_source_timeline_arc,
-    host_gpu_lim_register_texture, host_gpu_lim_resolve_texture_by_surface_id,
+    host_gpu_lim_clone_vertex_buffer, host_gpu_lim_command_queue,
+    host_gpu_lim_commit_and_wait_command_buffer, host_gpu_lim_commit_command_buffer,
+    host_gpu_lim_copy_pixel_buffer_to_texture, host_gpu_lim_copy_texture_command_buffer,
+    host_gpu_lim_create_command_buffer, host_gpu_lim_create_command_buffer_from_queue,
+    host_gpu_lim_drop_command_buffer, host_gpu_lim_drop_index_buffer,
+    host_gpu_lim_drop_pixel_buffer, host_gpu_lim_drop_pooled_texture_handle,
+    host_gpu_lim_drop_rhi_command_queue, host_gpu_lim_drop_storage_buffer, host_gpu_lim_drop_texture,
+    host_gpu_lim_drop_texture_registration, host_gpu_lim_drop_uniform_buffer,
+    host_gpu_lim_drop_vertex_buffer, host_gpu_lim_escalate_begin, host_gpu_lim_escalate_end,
+    host_gpu_lim_get_pixel_buffer, host_gpu_lim_host_video_source_timeline_arc,
+    host_gpu_lim_plane_base_address_pixel_buffer, host_gpu_lim_plane_size_pixel_buffer,
+    host_gpu_lim_register_texture, host_gpu_lim_resolve_pixel_buffer_by_surface_id,
+    host_gpu_lim_resolve_texture_by_surface_id,
     host_gpu_lim_resolve_texture_registration_by_surface_id,
-    host_gpu_lim_set_video_source_timeline_semaphore, host_gpu_lim_texture_native_dma_buf_fd,
+    host_gpu_lim_set_video_source_timeline_semaphore, host_gpu_lim_strong_count_pixel_buffer,
+    host_gpu_lim_surface_store, host_gpu_lim_texture_native_dma_buf_fd,
     host_gpu_lim_texture_registration_current_layout, host_gpu_lim_texture_registration_texture,
     host_gpu_lim_texture_registration_update_layout, host_gpu_lim_unregister_texture,
     host_gpu_lim_update_texture_registration_layout, host_gpu_lim_wait_timeline_semaphore,
 };
 use scope_token::with_full_scope_or_err;
-use shared::{handle_as_gpu_context, pixel_format_from_raw};
+use shared::pixel_format_from_raw;
 
 // pointers and reading nothing about layout.
 
@@ -108,927 +119,6 @@ unsafe extern "C" fn host_gpu_lim_drop_handle(owned_handle: *const c_void) {
     )
 }
 
-unsafe extern "C" fn host_gpu_lim_clone_pixel_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_pixel_buffer",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: `handle` is a `*const c_void` cast of
-            // `Arc::into_raw(Arc<PixelBufferRef>)` produced by
-            // `PixelBuffer::new` (host-side). Re-interpreting it as
-            // `*const PixelBufferRef` and bumping the strong count is the
-            // documented `Arc::increment_strong_count` contract.
-            unsafe {
-                Arc::increment_strong_count(handle as *const crate::core::rhi::PixelBufferRef);
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_pixel_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_pixel_buffer",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: matched with `host_gpu_lim_clone_pixel_buffer` and
-            // `PixelBuffer::new`'s `Arc::into_raw` initial bump.
-            // `Arc::decrement_strong_count` decrements; when refcount hits
-            // zero the underlying `PixelBufferRef` is dropped along with
-            // its platform buffer.
-            unsafe {
-                Arc::decrement_strong_count(handle as *const crate::core::rhi::PixelBufferRef);
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_strong_count_pixel_buffer(handle: *const c_void) -> usize {
-    run_host_extern_c(
-        "host_gpu_lim_strong_count_pixel_buffer",
-        || {
-            if handle.is_null() {
-                return 0;
-            }
-            // SAFETY: `handle` is `Arc::into_raw(Arc<PixelBufferRef>)`-shaped
-            // (see `PixelBuffer::new`'s `from_arc_into_raw`). We
-            // reconstruct the `Arc` temporarily, read the strong count, and
-            // immediately re-leak it via `Arc::into_raw` so the strong count
-            // returns to its pre-call value — `Arc::strong_count_from_raw`
-            // is not part of the public stable API. The reconstruction runs
-            // in HOST-COMPILED code regardless of caller DSO, so the cdylib
-            // never has to know `PixelBufferRef`'s in-memory layout.
-            unsafe {
-                let arc =
-                    Arc::from_raw(handle as *const crate::core::rhi::PixelBufferRef);
-                let count = Arc::strong_count(&arc);
-                let _ = Arc::into_raw(arc);
-                count
-            }
-        },
-        0,
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_plane_base_address_pixel_buffer(
-    handle: *const c_void,
-    plane_index: u32,
-) -> *mut u8 {
-    run_host_extern_c(
-        "host_gpu_lim_plane_base_address_pixel_buffer",
-        || {
-            if handle.is_null() {
-                return core::ptr::null_mut();
-            }
-            // SAFETY: `handle` is `Arc::into_raw(Arc<PixelBufferRef>)`-shaped;
-            // the leaked strong count keeps the `PixelBufferRef` alive for
-            // the duration of the call. We borrow `&PixelBufferRef` rather
-            // than reconstructing the Arc to avoid touching the refcount.
-            unsafe {
-                let pb_ref = &*(handle as *const crate::core::rhi::PixelBufferRef);
-                pb_ref.plane_base_address(plane_index)
-            }
-        },
-        core::ptr::null_mut(),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_plane_size_pixel_buffer(
-    handle: *const c_void,
-    plane_index: u32,
-) -> u64 {
-    run_host_extern_c(
-        "host_gpu_lim_plane_size_pixel_buffer",
-        || {
-            if handle.is_null() {
-                return 0;
-            }
-            // SAFETY: same as `host_gpu_lim_plane_base_address_pixel_buffer`.
-            unsafe {
-                let pb_ref = &*(handle as *const crate::core::rhi::PixelBufferRef);
-                pb_ref.plane_size(plane_index)
-            }
-        },
-        0,
-    )
-}
-// -------------------------------------------------------------------------
-// RhiCommandQueue Arc-handle lifecycle + create_command_buffer
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_clone_rhi_command_queue(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_rhi_command_queue",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: `handle` is `Arc::into_raw(Arc<RhiCommandQueueInner>)`-shaped.
-            unsafe {
-                Arc::increment_strong_count(
-                    handle as *const crate::core::rhi::command_queue::RhiCommandQueueInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_rhi_command_queue(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_rhi_command_queue",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: matched with the `Arc::into_raw` in
-            // `RhiCommandQueue::from_arc_into_raw`.
-            unsafe {
-                Arc::decrement_strong_count(
-                    handle as *const crate::core::rhi::command_queue::RhiCommandQueueInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_create_command_buffer_from_queue(
-    queue_handle: *const c_void,
-    out_cb: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_create_command_buffer_from_queue",
-        || -> i32 {
-            if queue_handle.is_null() {
-                write_err(
-                    "create_command_buffer_from_queue: null queue handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            if out_cb.is_null() {
-                write_err(
-                    "create_command_buffer_from_queue: null out_cb",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            // SAFETY: `queue_handle` is
-            // `Arc::into_raw(Arc<RhiCommandQueueInner>)`-shaped; the
-            // Arc's strong count keeps the inner alive for the duration.
-            let inner = unsafe {
-                &*(queue_handle
-                    as *const crate::core::rhi::command_queue::RhiCommandQueueInner)
-            };
-            let result = inner.inner.create_command_buffer();
-            match result {
-                Ok(platform_cb) => {
-                    let cb_inner =
-                        crate::core::rhi::command_buffer::CommandBufferInner {
-                            inner: platform_cb,
-                        };
-                    let cb = crate::core::rhi::CommandBuffer::from_inner(cb_inner);
-                    // SAFETY: out_cb points at caller-allocated stack
-                    // storage for a CommandBuffer value.
-                    unsafe {
-                        std::ptr::write(
-                            out_cb as *mut crate::core::rhi::CommandBuffer,
-                            cb,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-// -------------------------------------------------------------------------
-// CommandBuffer lifecycle: drop + consume-semantics commits (v7)
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_drop_command_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_command_buffer",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: matched with `Box::into_raw` in
-            // `CommandBuffer::from_inner`.
-            unsafe {
-                let _ = Box::from_raw(
-                    handle as *mut crate::core::rhi::command_buffer::CommandBufferInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_commit_command_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_commit_command_buffer",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: matched with `Box::into_raw` in
-            // `CommandBuffer::from_inner`; the cdylib's commit(self)
-            // nulls its local fields after this call so Drop won't
-            // double-free. We move-out of the Box so the platform
-            // commit can take ownership of the inner by-value.
-            let cb_box = unsafe {
-                Box::from_raw(
-                    handle as *mut crate::core::rhi::command_buffer::CommandBufferInner,
-                )
-            };
-            let cb_inner = *cb_box;
-            cb_inner.inner.commit();
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_commit_and_wait_command_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_commit_and_wait_command_buffer",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: see `host_gpu_lim_commit_command_buffer`.
-            let cb_box = unsafe {
-                Box::from_raw(
-                    handle as *mut crate::core::rhi::command_buffer::CommandBufferInner,
-                )
-            };
-            let cb_inner = *cb_box;
-            cb_inner.inner.commit_and_wait();
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_copy_texture_command_buffer(
-    handle: *const c_void,
-    src: *const c_void,
-    dst: *const c_void,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_copy_texture_command_buffer",
-        || {
-            if handle.is_null() || src.is_null() || dst.is_null() {
-                return;
-            }
-            // SAFETY: handle is `Box::into_raw(...)`-shaped; `&mut` is
-            // sound because the cdylib's `&mut self` guarantees no
-            // concurrent reference. src/dst are
-            // `*const Texture` (layout locked by `texture_layout` test).
-            unsafe {
-                let cb_inner = &mut *(handle
-                    as *mut crate::core::rhi::command_buffer::CommandBufferInner);
-                let src_tex = &*(src as *const crate::core::rhi::Texture);
-                let dst_tex = &*(dst as *const crate::core::rhi::Texture);
-                // Re-use the existing platform-specific copy_texture
-                // surface inside CommandBufferInner's `inner`.
-                #[cfg(all(
-                    not(feature = "backend-vulkan"),
-                    any(feature = "backend-metal", any(target_os = "macos", target_os = "ios"))
-                ))]
-                {
-                    cb_inner.inner.copy_texture(
-                        &src_tex.host_inner().inner,
-                        &dst_tex.host_inner().inner,
-                    );
-                }
-                #[cfg(any(
-                    feature = "backend-vulkan",
-                    all(target_os = "linux", not(feature = "backend-metal"))
-                ))]
-                {
-                    use crate::host_rhi::HostTextureExt;
-                    cb_inner
-                        .inner
-                        .copy_texture(src_tex.vulkan_inner(), dst_tex.vulkan_inner());
-                }
-                #[cfg(target_os = "windows")]
-                {
-                    cb_inner.inner.copy_texture(
-                        &src_tex.host_inner().inner,
-                        &dst_tex.host_inner().inner,
-                    );
-                }
-            }
-        },
-        (),
-    )
-}
-
-// -------------------------------------------------------------------------
-// GpuContextLimitedAccess command-queue / command-buffer / blit methods
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_command_queue(
-    gpu_handle: *const c_void,
-    out_queue: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_command_queue",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err("command_queue: null gpu handle", err_buf, err_buf_cap, err_len);
-                return 1;
-            };
-            if out_queue.is_null() {
-                write_err("command_queue: null out_queue", err_buf, err_buf_cap, err_len);
-                return 1;
-            }
-            // `gpu.command_queue()` returns `&RhiCommandQueue` (a borrow
-            // from GpuContext's stored field). Clone into a fresh owned
-            // β-shape for the caller — the Clone impl runs the host's
-            // `clone_rhi_command_queue` callback (Arc refcount bump).
-            let owned = gpu.command_queue().clone();
-            // SAFETY: out_queue points at caller-allocated stack storage.
-            unsafe {
-                std::ptr::write(
-                    out_queue as *mut crate::core::rhi::RhiCommandQueue,
-                    owned,
-                );
-            }
-            0
-        },
-        1,
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_create_command_buffer(
-    gpu_handle: *const c_void,
-    out_cb: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_create_command_buffer",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err(
-                    "create_command_buffer: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_cb.is_null() {
-                write_err(
-                    "create_command_buffer: null out_cb",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            match gpu.create_command_buffer() {
-                Ok(cb) => {
-                    // SAFETY: out_cb points at caller-allocated storage.
-                    unsafe {
-                        std::ptr::write(
-                            out_cb as *mut crate::core::rhi::CommandBuffer,
-                            cb,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(target_os = "linux")]
-unsafe extern "C" fn host_gpu_lim_copy_pixel_buffer_to_texture(
-    gpu_handle: *const c_void,
-    pixel_buffer: *const c_void,
-    texture: *const c_void,
-    surface_id_ptr: *const u8,
-    surface_id_len: usize,
-    width: u32,
-    height: u32,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_copy_pixel_buffer_to_texture",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err(
-                    "copy_pixel_buffer_to_texture: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if pixel_buffer.is_null() || texture.is_null() {
-                write_err(
-                    "copy_pixel_buffer_to_texture: null pixel_buffer or texture",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            // SAFETY: pixel_buffer / texture point at β-shape values
-            // whose layouts are locked by per-type regression tests.
-            let pb = unsafe { &*(pixel_buffer as *const crate::core::rhi::PixelBuffer) };
-            let tex = unsafe { &*(texture as *const crate::core::rhi::Texture) };
-            let id_bytes = unsafe { slice_from_raw(surface_id_ptr, surface_id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => {
-                    write_err(
-                        "copy_pixel_buffer_to_texture: surface_id not valid UTF-8",
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    return 1;
-                }
-            };
-            match gpu.copy_pixel_buffer_to_texture(pb, tex, id_str, width, height) {
-                Ok(()) => 0,
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(not(target_os = "linux"))]
-unsafe extern "C" fn host_gpu_lim_copy_pixel_buffer_to_texture(
-    _gpu_handle: *const c_void,
-    _pixel_buffer: *const c_void,
-    _texture: *const c_void,
-    _surface_id_ptr: *const u8,
-    _surface_id_len: usize,
-    _width: u32,
-    _height: u32,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    write_err(
-        "copy_pixel_buffer_to_texture: not available on this platform",
-        err_buf,
-        err_buf_cap,
-        err_len,
-    );
-    1
-}
-
-unsafe extern "C" fn host_gpu_lim_blit_copy(
-    gpu_handle: *const c_void,
-    src: *const c_void,
-    dst: *const c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_blit_copy",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err("blit_copy: null gpu handle", err_buf, err_buf_cap, err_len);
-                return 1;
-            };
-            if src.is_null() || dst.is_null() {
-                write_err("blit_copy: null src or dst", err_buf, err_buf_cap, err_len);
-                return 1;
-            }
-            // SAFETY: src / dst point at β-shape PixelBuffer values.
-            let src_pb = unsafe { &*(src as *const crate::core::rhi::PixelBuffer) };
-            let dst_pb = unsafe { &*(dst as *const crate::core::rhi::PixelBuffer) };
-            match gpu.blit_copy(src_pb, dst_pb) {
-                Ok(()) => 0,
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(target_os = "macos")]
-unsafe extern "C" fn host_gpu_lim_blit_copy_iosurface(
-    gpu_handle: *const c_void,
-    src_iosurface_ref: *const c_void,
-    dst_pixel_buffer: *const c_void,
-    width: u32,
-    height: u32,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_blit_copy_iosurface",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err(
-                    "blit_copy_iosurface: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if dst_pixel_buffer.is_null() {
-                write_err(
-                    "blit_copy_iosurface: null dst_pixel_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let dst_pb = unsafe {
-                &*(dst_pixel_buffer as *const crate::core::rhi::PixelBuffer)
-            };
-            let src_io = src_iosurface_ref as crate::apple::corevideo_ffi::IOSurfaceRef;
-            match unsafe { gpu.blit_copy_iosurface(src_io, dst_pb, width, height) } {
-                Ok(()) => 0,
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(not(target_os = "macos"))]
-unsafe extern "C" fn host_gpu_lim_blit_copy_iosurface(
-    _gpu_handle: *const c_void,
-    _src_iosurface_ref: *const c_void,
-    _dst_pixel_buffer: *const c_void,
-    _width: u32,
-    _height: u32,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    write_err(
-        "blit_copy_iosurface: not available on this platform (macOS-only)",
-        err_buf,
-        err_buf_cap,
-        err_len,
-    );
-    1
-}
-
-// -------------------------------------------------------------------------
-// GpuContextLimitedAccessVTable — surface_store accessors
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_surface_store(
-    gpu_handle: *const c_void,
-    out_store: *mut c_void,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_surface_store",
-        || {
-            // Always-clear: write a null-handle β-shape first so the
-            // caller has a defined state even on error paths.
-            if !out_store.is_null() {
-                unsafe {
-                    std::ptr::write(
-                        out_store as *mut crate::core::context::SurfaceStore,
-                        crate::core::context::SurfaceStore::null(),
-                    );
-                }
-            }
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                return;
-            };
-            if out_store.is_null() {
-                return;
-            }
-            // `gpu.surface_store()` returns `Option<SurfaceStore>` —
-            // a fresh β-shape with Arc refcount already bumped when
-            // Some. We write it into the out-param; the caller (cdylib
-            // or host) takes ownership.
-            if let Some(store) = gpu.surface_store() {
-                unsafe {
-                    std::ptr::write(
-                        out_store as *mut crate::core::context::SurfaceStore,
-                        store,
-                    );
-                }
-            }
-            // else: out_store already holds the null-handle β-shape.
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_check_out_surface(
-    gpu_handle: *const c_void,
-    id_ptr: *const u8,
-    id_len: usize,
-    out_pixel_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_check_out_surface",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err(
-                    "check_out_surface: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_pixel_buffer.is_null() {
-                write_err(
-                    "check_out_surface: null out_pixel_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let id_bytes = unsafe { slice_from_raw(id_ptr, id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => {
-                    write_err(
-                        "check_out_surface: surface_id not valid UTF-8",
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    return 1;
-                }
-            };
-            match gpu.check_out_surface(id_str) {
-                Ok(pb) => {
-                    unsafe {
-                        std::ptr::write(
-                            out_pixel_buffer as *mut crate::core::rhi::PixelBuffer,
-                            pb,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-
-// -------------------------------------------------------------------------
-// PixelBuffer acquire / get / resolve method-dispatch
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_acquire_pixel_buffer(
-    gpu_handle: *const c_void,
-    width: u32,
-    height: u32,
-    format_raw: u32,
-    out_pool_id_buf: *mut u8,
-    out_pool_id_cap: usize,
-    out_pool_id_len: *mut usize,
-    out_pixel_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_acquire_pixel_buffer",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err(
-                    "acquire_pixel_buffer: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_pixel_buffer.is_null() {
-                write_err(
-                    "acquire_pixel_buffer: null out_pixel_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let format = match pixel_format_from_raw(format_raw) {
-                Some(f) => f,
-                None => {
-                    let msg = format!(
-                        "acquire_pixel_buffer: invalid format_raw 0x{:08x}",
-                        format_raw
-                    );
-                    write_err(&msg, err_buf, err_buf_cap, err_len);
-                    return 1;
-                }
-            };
-            match gpu.acquire_pixel_buffer(width, height, format) {
-                Ok((pool_id, pb)) => {
-                    write_id_bytes(
-                        pool_id.as_str().as_bytes(),
-                        out_pool_id_buf,
-                        out_pool_id_cap,
-                        out_pool_id_len,
-                    );
-                    unsafe {
-                        std::ptr::write(
-                            out_pixel_buffer as *mut crate::core::rhi::PixelBuffer,
-                            pb,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_get_pixel_buffer(
-    gpu_handle: *const c_void,
-    pool_id_ptr: *const u8,
-    pool_id_len: usize,
-    out_pixel_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_get_pixel_buffer",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err(
-                    "get_pixel_buffer: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_pixel_buffer.is_null() {
-                write_err(
-                    "get_pixel_buffer: null out_pixel_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let id_bytes = unsafe { slice_from_raw(pool_id_ptr, pool_id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => {
-                    write_err(
-                        "get_pixel_buffer: pool_id not valid UTF-8",
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    return 1;
-                }
-            };
-            let pool_id = crate::core::rhi::PixelBufferPoolId::from_str(id_str);
-            match gpu.get_pixel_buffer(&pool_id) {
-                Ok(pb) => {
-                    unsafe {
-                        std::ptr::write(
-                            out_pixel_buffer as *mut crate::core::rhi::PixelBuffer,
-                            pb,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_resolve_pixel_buffer_by_surface_id(
-    gpu_handle: *const c_void,
-    surface_id_ptr: *const u8,
-    surface_id_len: usize,
-    out_pixel_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_resolve_pixel_buffer_by_surface_id",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(gpu_handle) }) else {
-                write_err(
-                    "resolve_pixel_buffer_by_surface_id: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_pixel_buffer.is_null() {
-                write_err(
-                    "resolve_pixel_buffer_by_surface_id: null out_pixel_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let id_bytes = unsafe { slice_from_raw(surface_id_ptr, surface_id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => {
-                    write_err(
-                        "resolve_pixel_buffer_by_surface_id: surface_id not valid UTF-8",
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    return 1;
-                }
-            };
-            match gpu.resolve_pixel_buffer_by_surface_id(id_str) {
-                Ok(pb) => {
-                    unsafe {
-                        std::ptr::write(
-                            out_pixel_buffer as *mut crate::core::rhi::PixelBuffer,
-                            pb,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
 /// Static [`GpuContextLimitedAccessVTable`] installed once per process.
 /// Paired with the per-RuntimeContext gpu-limited handle returned by
 /// [`HOST_RUNTIME_CONTEXT_VTABLE`]`::gpu_limited_access`.
